marshal: rename marshal to serialization
