+++ serialization/counter/marshal.go
