+++ serialization/smallint/marshal.go
