+++ tests/serialization/marshal_2_tinyint_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/gocql/gocql"
 	"github.com/gocql/gocql/internal/tests/serialization"
 	"github.com/gocql/gocql/internal/tests/serialization/mod"
-	"github.com/gocql/gocql/marshal/tinyint"
+	"github.com/gocql/gocql/serialization/tinyint"
 )
 
 func TestMarshalTinyint(t *testing.T) {
