+++ tests/serialization/marshal_5_bigint_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/gocql/gocql"
 	"github.com/gocql/gocql/internal/tests/serialization"
 	"github.com/gocql/gocql/internal/tests/serialization/mod"
-	"github.com/gocql/gocql/marshal/bigint"
+	"github.com/gocql/gocql/serialization/bigint"
 )
 
 func TestMarshalBigInt(t *testing.T) {
