+++ tests/serialization/marshal_6_counter_corrupt_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/gocql/gocql"
 	"github.com/gocql/gocql/internal/tests/serialization"
 	"github.com/gocql/gocql/internal/tests/serialization/mod"
-	"github.com/gocql/gocql/marshal/counter"
+	"github.com/gocql/gocql/serialization/counter"
 )
 
 func TestMarshalCounterCorrupt(t *testing.T) {
